Add visitor/fields
